refactor(runtime): split timer drivers + fix RwLock/Mutex poison handling
Review findings addressed:
- #1: split timer.rs (959 lines) — ThreadDriver, AppDriver, PoolDriver moved to timer_drivers.rs (major)
- #2: connection_table.rs RwLock::write/read().unwrap() → .unwrap_or_else(|e| e.into_inner()) across register, remove, remove_by_receiver, receivers_for_signal (major)
- #3: event_loop.rs run() Mutex::lock().unwrap() → .expect("reason") (minor)
- #4: thread_pool.rs worker Mutex::lock().unwrap() → .expect("reason") (minor)
- #5: remove stale # Panics sections from connection_table.rs methods (nit)
- #8: timer.rs background callback silent lock failure → .unwrap_or_else(|e| e.into_inner()) (minor)

Deferred:
- #6: meta.rs soft-limit (846 lines) — counter-rule: cohesive metadata types, no split needed
- #7: signal.rs emit SingleShot retain-pass — informational, no defect
- #9: application.rs try_with_tree — already correct
- #10: connect_typed debug_assert release behavior — already documented in # Panics
